Видалив зайві файли тестів
